Chatbot_3.6/requirements1.txt
aiml
lxml
beautifulsoup4
flask
requests
wikipedia
numpy
configparser
wordsegment
networkx
mysql-connector-python
nltk
stanfordcorenlp
future
pyaudio
pyttsx3
google-cloud-core
tqdm
scikit-learn
tensorlayer==1.6.3
tensorflow==1.4

# python -m nltk.downloader punkt